requirements.txt
opencv-python
mss
pyautogui
pyyaml
requests
Pillow

